requirements.txt
# Core video processing
moviepy==1.0.3
ffmpeg-python==0.2.0

# Audio separation
demucs==4.0.1

# Transcription
openai-whisper==20231117
torch==2.3.0
torchaudio==2.3.0

# AI APIs
anthropic==0.28.0

# Utilities
python-dotenv==1.0.1
rich==13.7.1
psutil==5.9.8
Pillow==10.3.0
numpy==1.26.4
